Create COSW10-hw-Hobbies
This homework is going to be a little bit of a challenge, but I am confident that you got this. The instructions are a little vague, the intent is for you to refer to the example and get as close to it as you can. Use CSS to create margins, padding and alignment. Use semantic divs for the sections (header, main, footer). Include 2 Google fonts in your webpage.  Add the basic information for the footer too.