Create COSW-week1-hw
For this weeks homework you will be creating a new page in CodeSandbox. Review the image below and try your best to duplicate the page layout shown. The red arrows and text are there as guides and not something that needs to be included in your assignment. If you are unable to see the image, please let me know immediately. Please name the page with the following naming convention: "COSW10 - Week 1 - Homework". Feel free to add any addition info like, year/term/semester.