Create Week 3: hw
In my work as a front-end web developer, it is very common for me to receive Photoshop or image files from designers for new features or layouts on the website. I will look at the specifications and images they have provided and convert that into HTML and CSS code.

For this activity you have been hypothetically tasked by Long Beach City College to create a landing page for the Web Development associate's degree. They have provided a basic mock of what they want the layout to be but, like many free lance projects, they want you to just pull in content and images off of their current websites. This assignment is just for a single web page, but there will be "links" to other pages that you will need to just add to the page (you can link to whatever you think would be relevant).