Create CSS Transitions
Create a webpage on a topic of your choice with the following elements:

Do a transition or animation for an h1 heading at the top of your page.
Find 3 images of the same size (and theme) and display on the page. Do a hover filter on the images - choose three different filters to apply. Visit W3School's CSS Filter pageLinks to an external site..
Using W3School's CSS Image OverlaysLinks to an external site., pick 1 overlay to add to the one of your images.
Embed at least one video into your webpage.
Create an accordion FAQ page with 3 questions/answers of your own.